# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet EDR API
  version: "0.1.0"
  description: |
    HTTP API for Fleet EDR. Three audience tiers:

    - **Agents** (macOS endpoints) authenticate with a per-host bearer
      token obtained at enrollment. They call `POST /api/enroll`,
      `POST /api/events`, and the host-scoped `/commands` endpoints.
    - **Browsers** (admin UI) authenticate with an `edr_session` cookie
      (HttpOnly, Secure, SameSite=Lax) plus an `X-CSRF-Token` header on
      unsafe methods. They call the session-protected endpoints.
    - **Operators** hit `/livez`, `/readyz`, `/health` with no auth.

    For prose context see
    [docs/api.md](https://github.com/getvictor/fleet-edr/blob/main/docs/api.md).
    For a human-readable host/agent install guide see
    [docs/install-agent-manual.md](https://github.com/getvictor/fleet-edr/blob/main/docs/install-agent-manual.md).
  license:
    name: Fleet EDR License
    url: https://github.com/getvictor/fleet-edr/blob/main/LICENSE
servers:
  - url: https://edr.example.com
    description: Your EDR server
...
paths:
...
  /api/hosts/{host_id}/tree:
    get:
      tags: [Hosts]
      summary: Process tree for a host
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: from
          in: query
          schema: {type: integer, format: int64}
          description: Lower bound, ns since epoch. Default `now - 1h`.
        - name: to
          in: query
          schema: {type: integer, format: int64}
          description: Upper bound, ns since epoch. Default `now`.
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 5000, default: 2000}
      responses:
        "200":
          description: Forest of process roots
          content:
            application/json:
              schema:
                type: object
                properties:
                  roots:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProcessNode"
...
components:
  securitySchemes:
    HostToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Per-host bearer token issued by `POST /api/enroll`. Pass as
        `Authorization: Bearer <host_token>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: edr_session
      description: |
        HttpOnly cookie set by `POST /api/session`. The browser
        attaches it automatically.
    CSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on unsafe methods (`POST`, `PUT`, `DELETE`) for
        session-cookie-authenticated endpoints. Value comes from the
        login response's `csrf_token` field.
  parameters:
    HostID:
      name: host_id
      in: path
      required: true
      description: Host identifier (UUID-shaped opaque string)
      schema: {type: string}
  schemas:
...
    Process:
      type: object
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        pid: {type: integer}
        ppid: {type: integer}
        path: {type: string}
        args: {type: array, items: {type: string}}
        uid: {type: integer}
        gid: {type: integer}
        code_signing:
          type: object
          properties:
            team_id: {type: string}
            signing_id: {type: string}
            flags: {type: integer}
            is_platform_binary: {type: boolean}
        sha256: {type: string}
        fork_time_ns: {type: integer, format: int64}
        exec_time_ns: {type: integer, format: int64}
        exit_time_ns: {type: integer, format: int64}
        exit_code: {type: integer}
    ProcessNode:
      allOf:
        - $ref: "#/components/schemas/Process"
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: "#/components/schemas/ProcessNode"
...